+++ app/src/main/java/com/example/gsyvideoplayer/DetailNormalActivityPlayer.java
@@ -129,7 +129,7 @@ public class DetailNormalActivityPlayer extends Activity {
             })
             .setGSYVideoProgressListener(new GSYVideoProgressListener() {
                 @Override
-                public void onProgress(int progress, int secProgress, int currentPosition, int duration) {
+                public void onProgress(long progress, long secProgress, long currentPosition, long duration) {
                     Debuger.printfLog(" progress " + progress + " secProgress " + secProgress + " currentPosition " + currentPosition + " duration " + duration);
                 }
             })
